Extract apply manifest logic in common Remove redundant code
